+++ client/src/components/App.js
@@ -2,13 +2,6 @@ import React, { Component } from "react";
 import { StripeProvider } from "react-stripe-elements";
 import { BrowserRouter as Router } from "react-router-dom";
 import axios from "axios";
-import {
-  AboutYou,
-  OtherInformation,
-  AboutYouDetails,
-  OtherInformationDetails,
-  Demographic
-} from "./Common/ProfileComponents";
 
 // Antd style
 import "antd/lib/layout/style/index.css";
@@ -107,14 +100,7 @@ class App extends Component {
 
     return (
       <StripeProvider stripe={stripe}>
-        <div style={{ margin: "5rem" }}>
-          <AboutYou />
-          <OtherInformation />
-          <AboutYouDetails />
-          <OtherInformationDetails />
-          <Demographic />
-        </div>
-        {/* <Router>
+        <Router>
           <div className="App">
             <Navbar
               isLoggedIn={isLoggedIn}
@@ -127,7 +113,7 @@ class App extends Component {
               {...this.state}
             />
           </div>
-        </Router> */}
+        </Router>
       </StripeProvider>
     );
   }
