+++ client/src/components/Common/ProfileComponents/Demographic/index.js
@@ -6,7 +6,7 @@ import { Row, Col } from "antd";
 import Field from "../Field";
 import fields from "../../../../constants/fields";
 
-const Demographic = ({ errors = {}, handleError }) => {
+const Demographic = ({ data = {}, errors = {}, handleChange, handleError }) => {
   return (
     <SectionWrapper>
       <Title
@@ -18,32 +18,53 @@ const Demographic = ({ errors = {}, handleError }) => {
           <Col xs={24} sm={12}>
             <Field
               {...fields["sexualOrientation"]}
+              value={data["sexualOrientation"]}
               error={errors["sexualOrientation"]}
+              handleChange={handleChange}
             />
           </Col>
           <Col xs={24} sm={12}>
-            <Field {...fields["degreeLevel"]} error={errors["degreeLevel"]} />
+            <Field
+              {...fields["degreeLevel"]}
+              value={data["degreeLevel"]}
+              error={errors["degreeLevel"]}
+              handleChange={handleChange}
+            />
           </Col>
         </Row>
         <Row gutter={25}>
           <Col xs={24} sm={12}>
-            <Field {...fields["ethnicity"]} error={errors["ethnicity"]} />
+            <Field
+              {...fields["ethnicity"]}
+              value={data["ethnicity"]}
+              error={errors["ethnicity"]}
+              handleChange={handleChange}
+            />
           </Col>
           <Col xs={24} sm={12}>
             <Field
               {...fields["earningOfParents"]}
+              value={data["earningOfParents"]}
               error={errors["earningOfParents"]}
+              handleChange={handleChange}
             />
           </Col>
         </Row>
         <Row gutter={25}>
           <Col xs={24} sm={12}>
-            <Field {...fields["disability"]} error={errors["disability"]} />
+            <Field
+              {...fields["disability"]}
+              value={data["disability"]}
+              error={errors["disability"]}
+              handleChange={handleChange}
+            />
           </Col>
           <Col xs={24} sm={12}>
             <Field
               {...fields["parentsWorkInPress"]}
+              value={data["parentsWorkInPress"]}
               error={errors["parentsWorkInPress"]}
+              handleChange={handleChange}
             />
           </Col>
         </Row>
@@ -52,14 +73,21 @@ const Demographic = ({ errors = {}, handleError }) => {
           <Col xs={24} sm={12}>
             <Field
               {...fields["caringResponsibilities"]}
+              value={data["caringResponsibilities"]}
               error={errors["caringResponsibilities"]}
+              handleChange={handleChange}
             />
           </Col>
         </Row>
 
         <Row gutter={25}>
           <Col xs={24} sm={12}>
-            <Field {...fields["allergies"]} error={errors["allergies"]} />
+            <Field
+              {...fields["allergies"]}
+              value={data["allergies"]}
+              error={errors["allergies"]}
+              handleChange={handleChange}
+            />
           </Col>
         </Row>
 
@@ -67,7 +95,9 @@ const Demographic = ({ errors = {}, handleError }) => {
           <Col xs={24} sm={12}>
             <Field
               {...fields["consentedOnPressPadTerms"]}
+              value={data["consentedOnPressPadTerms"]}
               error={errors["consentedOnPressPadTerms"]}
+              handleChange={handleChange}
             />
           </Col>
         </Row>
