+++ client/src/components/Common/TabbedView/TabbedView.style.js
@@ -6,14 +6,16 @@ export const CardContainer = styled.div`
   }
 
   .ant-tabs .ant-tabs-top-content > .ant-tabs-tabpane {
-    box-shadow: 0px 0px 24px rgba(0, 0, 0, 0.04);
-    padding: 1rem;
+    box-shadow: 0px 0px 13px 1px rgba(0, 0, 0, 0.4);
+    width: 97%;
+    margin: 0 1rem 1rem;
   }
 
   .ant-tabs.ant-tabs-card .ant-tabs-card-bar .ant-tabs-tab {
     color: ${colors.fontPrimary};
     opacity: 0.5;
-    margin: 0 5px;
+    margin-right: 5px;
+    box-shadow: 0px 0px 13px 1px rgba(0, 0, 0, 0.75);
   }
 
   .ant-tabs.ant-tabs-card .ant-tabs-card-bar .ant-tabs-tab-active {
